Bared Fixes

+++ system/application/controllers/cart.php
@@ -27,13 +27,13 @@ class Cart extends Controller {
 	
 	function check_session()
 	{
-		/*if($this->session->userdata('previous'))
+		if($this->session->userdata('previous'))
 		{
 			if(basename($_SERVER['PHP_SELF']) != $this->session->userdata('previous'))
 			{
 				$this->session->destroy();
 			}
-		}*/
+		}
 	}
 		
 	
@@ -152,10 +152,15 @@ class Cart extends Controller {
 		$prods=array();
 		$cats=array();
 		$gift_card=0;
+		$only_gift_voucher = true;
+
 		foreach($cart as $item) 
 		{
 			if($item['product_id']==11){$gift_card=$gift_card+$item['price'];}
 			else{
+				# stupid way to code really - this is band aid solution to ignore shipping if cart only has gift voucher.
+				$only_gift_voucher = false;
+			
 				$product = $this->Product_model->identify($item['product_id']);
 				$total_weight=$total_weight+($product['weight']*$item['quantity']);
 				$total_volume=$total_volume+($product['volume']*$item['quantity']);
@@ -677,7 +682,8 @@ class Cart extends Controller {
 		$shipping_weight_method = 0;
 		$shipping_cost = 0;
 		if(!$this->session->userdata('store_pickup')){
-
+			# if cart has other items other than gift voucer
+			if(!$only_gift_voucher){
 			$method = $this->System_model->get_shipping($shipping_method);
 			if($method['price_type']==1){
 				$cost_weight=$method['price_value'];
@@ -713,6 +719,7 @@ class Cart extends Controller {
 					$shipping_cost = $condition['price'];
 				}
 			}
+			} # end gift voucher check
 		}
 		
 		$total = $subtotal2 + $shipping_cost + $gift_card;
@@ -3584,11 +3591,9 @@ class Cart extends Controller {
 
 	function send_giftcard()
 	{
-		
-		
 		$ndate = date('Y-m-d');
 		$list = $this->Cart_model->get_giftcard($ndate);
-		
+	
 		foreach($list as $l)
 		{
 			$product = $this->Product_model->identify($l['product_id']);
@@ -3631,7 +3636,7 @@ class Cart extends Controller {
 								<span style="font-weight: 700; font-size: 18px;">AU $ '.number_format($l['price'],2,'.',',').'</span><br/>
 								<br/>
 								ENTER CODE AT CHECKOUT<br/>
-								<span style="font-weight: 700; font-size: 18px;">'.$l['gift_code'].'</span><br/>
+								<span style="font-weight: 700; font-size: 18px; text-transform: none;">'.$l['gift_code'].'</span><br/>
 								<br/>
 								VALID UNTIL<br/>
 								<span style="font-weight: 700; font-size: 18px;">'.$to_date.'</span>
@@ -3644,27 +3649,27 @@ class Cart extends Controller {
 						<td>
 							Dear '.$l['gift_to'].',<br/>
 							<br/>
-							'.$l['gift_from'].' has sent you a Gift Voucher to spend on something special from Bared Footware. Your Gift Voucher can be redeemed online at <a href="'.base_url().'">www.bared.com.au</a> and is valid for 12 months.<br/>
+							'.$l['gift_from'].' has sent you a gift voucher to spend on something special from Bared Footwear. Your gift voucher can be redeemed in-store, online at <a href="'.base_url().'">www.bared.com.au</a> or by calling us on (03) 9509 5771.<br/>
 							<br/>
 							Now you just need to choose your gift, wait for your delivery, and enjoy!<br/><br/>
-							xo Bared Footware<br/>
+							xo Bared Footwear<br/>
 							<br/>
 							<br/>
 							<span style="font-weight: 700">HOW TO REDEEM YOUR GIFT VOUCHER</span><br/>
-							Your Gift Voucher has a unique code, which is required to ‘pay’ for your item. <br/>
+							Your gift voucher has a unique code, which is required to \'pay\' for your item. <br/>
 							<ol>
 								<li>Go to <a href="'.base_url().'">www.bared.com.au</a></li>
 								<li>Select the item you wish to purchase.</li>
 								<li>Add to Shopping Bag.</li>
-								<li>To purchase, go to your Shopping Bag, located in the top right of the screen, and follow the prompts. You will be required to sign in or register with us, and we will also ask you for a ‘Billing address’, even though you may not be required to pay any extra for your item. If there is an outstanding balance, you will be required to give your credit card details also.</li>
-								<li>When you get to the ORDER SUMMARY page, you will need to add your Gift Voucher Code into the area allocated and select UPDATE.</li>
+								<li>To purchase, go to your Shopping Bag, located in the top right of the screen, and follow the prompts. You will be required to sign in or register with us, and we will also ask you for a \'billing address\', even though you may not be required to pay any extra for your item. If there is an outstanding balance, you will be required to give your credit card details also.</li>
+								<li>When you get to the ORDER SUMMARY page, you will need to add your gift voucher code into the area allocated and select UPDATE.</li>
 								<li>Follow the prompts through to place your order.</li>
 								<li>Wait for your delivery to arrive!</li>
 							</ol>
 							<br/>
-							If there is a difference between the amount on the Gift Voucher and the order total you may need to add your credit card details to pay any extra amount.<br/><br/>
-							If there is an amount left over, it will remain available on your Gift Voucher for future use.<br/><br/>
-							Gift Vouchers are valid for 12 months from the date of purchase and can only be used in the Bared Footware.<br/>
+							If there is a difference between the gift voucher amount and the order total you\'ll need to enter additional payment details (credit card or PayPal).<br/><br/>
+							If there is an amount left over, it will remain available on your gift voucher for future use.<br/><br/>
+							Gift vouchers are valid for 12 months from the date of purchase and can only be used in the Bared Footwear.<br/>
 						</td>
 					</tr>
 				</table>
@@ -3684,16 +3689,16 @@ class Cart extends Controller {
 					</tr>
 				</table>';
 				
-			//echo $msg;
+			#echo $msg;exit;
 			
 			$this->load->library('email');
 			$config['mailtype'] = 'html';	
 			$this->email->initialize($config);	
-			$this->email->from('noreply@bared.com.au','Bared Footware');
+			$this->email->from('noreply@bared.com.au','Bared Footwear');
 			$this->email->to($l['gift_email']);
-			$this->email->bcc('raquel@propagate.com.au');
+			#$this->email->bcc('raquel@propagate.com.au');
 			
-			$this->email->subject('Bared Footware - Gift Voucher @ Bared Footware');
+			$this->email->subject('Bared Footwear - Gift Voucher @ Bared Footwear');
 			$this->email->message($msg);
 			$sent = $this->email->send();
 		}
@@ -5449,7 +5454,7 @@ class Cart extends Controller {
 		$header .='
 		<link href="http://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet" type="text/css">
 		<table width="630" align="center"><tr><td></td></tr>
-		<tr>
+			<tr>
 				<td>
 				<span style="font-family:Arial, sans-serif; font-size:18px; font-weight:400; ">
 				Thank you for shopping with Bared. You should receive your shoes within four to eight working days (or often sooner!), and we hope you really love them.<br><br>
@@ -5458,13 +5463,11 @@ class Cart extends Controller {
 				
 				We\'ll happily send you another pair (or two!), at no cost to you, to help you compare sizes or styles, and we\'ll pay the return postage for any or all pairs if they\'re still not quite right.<br><br>
 				
-				Your\'s,<br> 
+				Yours,<br> 
 				The Bared Team
 				</span>
 				</td>
-				</tr>
-				<tr><td>&nbsp;</td></tr>
-		';
+			</tr><tr><td>&nbsp;</td></tr>';
 		$footer ='';
 		$footer .='
 				<table width="630" align="center">
@@ -5487,7 +5490,8 @@ class Cart extends Controller {
 				<tr>
 				<td align="center"> 
 				<br>
-				<span style="font-family:Times New Roman, Times, serif; font-size:14px; font-weight:700; color:#000!important; text-decoration:none; ">WWW.BARED.COM.AU </span>
+				<span style="font-family:Times New Roman, Times, serif; font-size:14px; font-weight:700; color:#000!important; text-decoration:none; ">WWW.BARED.COM.AU </span><br> 
+				</td>
 		</tr></table>';
 		$content='';
 		if(!$order['store_pickup']){
@@ -5696,7 +5700,6 @@ class Cart extends Controller {
 		</table>
 		';			
 		$message = $header.$content.$footer;
-		
 		#echo $message;exit;
 		
 		$this->load->library('email');
